update default type for beta

+++ src/python/impactx/dashboard/Input/defaults.py
@@ -145,7 +145,7 @@ class DashboardDefaults:
         "mass_MeV": "float",
         "charge_qe": "int",
         "csr_bins": "int",
-        "beta": "int",
+        "beta": "float",
         "emitt": "float",
         "alpha": "float",
     }
